Poverty And Unemployment
Problem Statement - Poverty is rooted in unemployment and underemployment. For the impoverished, labour is sometimes their sole means of improving their situation. As a result, creating productive job opportunities is critical for poverty reduction and long-term economic and social development. It is critical to offer respectable employment for the underprivileged, particularly women and young people, who give combined income and strength. Job possibilities are most likely a big component that has arisen in Pakistan recently. We all know, there are a number of things that have an influence on the economy. Inflation rises as a result of imports and exports going down, economic and non-economic growth, food costs increase, health, education, and other variables. If we see on another side inflation also affects the education index of the country. Rapid economic growth has the potential to result in a rapid increase in productive and remunerative jobs, reducing poverty. Nonetheless, the growth process' contributions to poverty reduction is dependent not only on the pace of economic progress, but also from the poor's capacity to adapt to the rising demand for labour in more productive occupations. A social perspective on development stresses the belief that good labour is the greatest path to socioeconomic progress, poverty elimination, and personal happiness. Job creation will play a significant role in reaching internationally agreed development targets, particularly the Millennium Development Goal of reducing severe poverty by 2015. To analyze all the factors in Pakistan and other countries, the unlimited independent variables that can influence education, unemployment and inflation. Some of them that I used to build the model are as MICROSOFT EXCEL FILE. The dataset collects from (INDICATOR INDEX, N.D.) feature indicators used by different factors (In every indicator there are a number of series available and choose only valuable feature that can give better predictions). For Analysis I use regression models.